Fixed fish disappearing/font not rendering correct
Resolved a memory leak that caused the watch face to allocate an infinite number of fishes, causing the fishes to die. Fishes now are only birthed once, and live forever.
Resolved a similar issue with the font. Font was being loaded every second, and this caused issues with the font not loading correctly.